FileUpload/requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
Deprecated==1.2.13
Django==4.0.2
django-redis==5.2.0
et-xmlfile==1.1.0
faiss-cpu==1.7.2
h11==0.13.0
idna==3.3
numpy==1.22.2
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
Pillow==9.0.1
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
redis==4.1.4
selenium==4.1.2
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
torch==1.10.2
torchvision==0.11.3
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
wrapt==1.13.3
wsproto==1.1.0
